refactored server code to separate index file
