+++ server/index.js
@@ -1,14 +1,6 @@
-const express = require('express');
-const app = express();
+const app = require('./app.js');
 const port = 3003;
-const bodyParser = require('body-parser');
 
-
-app.use(express.static(__dirname + '../client/dist'));
-app.use(bodyParser.json());
-
-app.get('/', (req, res) => {
-  res.send('Hello World');
-}); 
-
-app.listen(port, () => console.log(`app is listening on port ${port}`));
+app.listen(port, () => {
+  console.log(`app is listening on port ${port}`);
+});
